WIP: now the partials that are computed are not controlled by the orchestrator, he only receives what the backends produced as partials, according to their own `storeInterval` configurations. They will therefore be able to take more or less snapshots depending on the size of stores, or other heuristics.
The orchestrator should therefore only delete partials when they are off-bounds
in its perspective, and that they are the _last_ partial required to fulfill
the requested start-block (like 80-92), otherwise, it flushes a snapshot. It
should not itself attempt to schedule jobs off bounds of its perceived
`storeInterval` and will therefore Squash at its perceived `storeInterval`,
at the end of each subrequests. This isn't perfect, but allows the subrequests
to be more or less adaptative.

reqChunks can now be only ranges.. as they are not responsible for deciding
what _partials_ should be produced _nor_ expected by the Squasher to be
produced.  They will receive it through the Trailer of the subrequest.