Added some notes and comments on the patterns, things remaining, things to adjust.
Still missing: ensuring this all works with chains that skip blocks. There
are a few notes to that effect.. but basically, we'll need, in the pipeline
to figure out when we cross `storeSaveInterval` boundaries, to get the boundary number
and flush a snapshot/partial before actually processing the next block.
All boundaries of stores need to be programmatic like that, and not depend
_directly_ on the blocks from the pipeline. A block _equal of after_ such
a boundary should trigger those saving events.