Removed `pipeline.RequestContext` stack recording of span(s)
This was error prone and required extra tracking elements that add little benefits. Most of the places where the span was needed, it was nearby (in callsite) and just passing around the `span` is a much safe approach.

The end span has been refactored also to also call `RecordError` when the span is an error.