Removed specialized `errors.GRPCError` and uses standard `status.Error` everywhere
Extraction of any `status.Error` in the chain is done automatically while retaining full error cause for addition in the tracing span.

Moved error handling out of `StreamEndedWithErr` closer to `service` and renamed the method to `OnStreamTerminated` with a godoc explaining what it does. I renamed it because it does not actually handle non-gracefull `err` anymore.