Reader tests fixed. Caused by the mis-used of `append()`
See: https://trstringer.com/golang-append/#:~:text=The%20append%20built%2Din%20function,Append%20returns%20the%20updated%20slice.
and: https://go.dev/doc/effective_go#append

Basically, because the length of the system protos had changed (because of the removal
of test), we triggered a mis-aligned slice size, and we mutated the
`systemDefs` variable instead of returning a new slice, or something like that.