Fixed a bug where the server could "jump" to live in a stream
This fixes a bug hit by a user in production. The problem is the following:
- On request, we starts `InitAndBackprocess` which is blocking
- The `InitAndBackprocess` starts the exec output reader code and blocks on it waiting for it to terminate sending all existing blocks to user
- The blocking is unblocked on 2 conditions, if the reader terminated or if the context of the call is canceled.

The problem arose when the server receives the shutdown signal but is still waiting on some internal work to complete (for example the recently metrics handling is now waited on to complete).

When this happen, the context is canceled which unblocks the blocked code which ultimately unblocks `InitAndBackprocess`. The unblocking when `ctx.Done()` is hit however returns without error nor anything else that would lead the caller to know that the exec out reader did not actually terminated but simply stopped.

Now, the request flow continue and after `InitAndBackprocess`, we are starting to send blocks after the linear handoff. Which means that if there is enough time between the server's received Ctrl-C signal and the actual shutdown, "live" blocks will be sent to the stream and ultimately, it will lead to all sort of problems in the consumer side.

The fix is simply to return an error when the context is canceled, that way, the caller will not do anything and will stop the request with a gRPC canceled code.