Updated Readme by removing the link
There is no need to have an "Additional Information"  section in the "https://substreams.streamingfast.io/developers-guide/modules" page because the link added to this section itself leads to this "https://substreams.streamingfast.io/developers-guide/modules" or the same page. Need to remove it for now or we need to link it with other sources if we have one.