This enables running module with unknown imports as long as it's not called
We achieve this my binding every unknown module and their imports to a dummy implementation that fails as soon as it's called.

This enables Substreams to import for example `alloy` in the Ethereum world to perform some Ethereum related operations. It also enables developer to pull in `solana_program` in Solana world so they are able to decode IDL directly.