Added status bar, rework of the Output panel, which is now the first to be shown. Backprocessing is now second.
Request panel needs to be resurrected, probably with a form, and then
all the logic that lives in `gui.go` needs to be moved to create a new
Request _in here_, rather than use the params from the CLI always. We'll
want to recreate a new Request, based on the `Config` that we received, but tweaked.