Remove enforcement of `BUFBUILD_AUTH_TOKEN`
It appears a Buf token is not mandatory anymore to fetch descriptors set from Buf API anymore. There is probably rate limits but outside of this, seems that it's possible to fetch it without a token.

If we see that we start to have some rate limit errors, we shall re-introduced a specific handling to ask use to provide a `BUFBUILD_AUTH_TOKEN`.